requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
opencv-python-headless
dlib
face_recognition
torch
transformers
sentencepiece
scikit-learn
faiss-cpu
git+https://github.com/openai/CLIP.git
